Language.PlutusIR

Documentation

newtype TyName a

We use a newtype to enforce separation between names used for types and those used for terms.

Constructors

TyName

Fields

unTyName :: Name a

data Name a

A Name represents variables/names in Plutus Core.

Constructors

Name

Fields

nameAttribute :: a

nameString :: Text

The identifier name, for use in error messages.

nameUnique :: Unique

A Unique assigned to the name during lexing, allowing for cheap comparisons in the compiler.

data VarDecl (tyname :: * -> *) (name :: * -> *) a

A "variable declaration", i.e. a name and a type for a variable.

Constructors

VarDecl

Fields

varDeclAnn :: a

varDeclName :: name a

varDeclType :: Type tyname a

data TyVarDecl (tyname :: * -> *) a

A "type variable declaration", i.e. a name and a kind for a type variable.

Constructors

TyVarDecl

Fields

tyVarDeclAnn :: a

tyVarDeclName :: tyname a

tyVarDeclKind :: Kind a

varDeclNameString :: VarDecl name Name a -> String

tyVarDeclNameString :: TyVarDecl TyName a -> String

data Kind a

Kinds. Each type has an associated kind.

Constructors

Type a

KindArrow a (Kind a) (Kind a)

Size a

data Type (tyname :: * -> *) a

A Type assigned to expressions.

Constructors

TyVar a (tyname a)

TyFun a (Type tyname a) (Type tyname a)

TyIFix a (Type tyname a) (Type tyname a)
Fix-point type, for constructing self-recursive types

TyForall a (tyname a) (Kind a) (Type tyname a)

TyBuiltin a TypeBuiltin
Builtin type

TyInt a Natural
Type-level size

TyLam a (tyname a) (Kind a) (Type tyname a)

TyApp a (Type tyname a) (Type tyname a)

data Datatype tyname name a

Constructors

Datatype a (TyVarDecl tyname a) [TyVarDecl tyname a] (name a) [VarDecl tyname name a]

datatypeNameString :: Datatype TyName name a -> String

data Recursivity

Constructors

NonRec

Rec

data Binding tyname name a

Constructors

TermBind a (VarDecl tyname name a) (Term tyname name a)

TypeBind a (TyVarDecl tyname a) (Type tyname a)

DatatypeBind a (Datatype tyname name a)

data Term tyname name a

Constructors

Let a Recursivity [Binding tyname name a] (Term tyname name a)

Var a (name a)

TyAbs a (tyname a) (Kind a) (Term tyname name a)

LamAbs a (name a) (Type tyname a) (Term tyname name a)

Apply a (Term tyname name a) (Term tyname name a)

Constant a (Constant a)

Builtin a (Builtin a)

TyInst a (Term tyname name a) (Type tyname a)

Error a (Type tyname a)

IWrap a (Type tyname a) (Type tyname a) (Term tyname name a)

Unwrap a (Term tyname name a)

data Program tyname name a

Constructors

Program a (Term tyname name a)

embedIntoIR :: Term tyname name a -> Term tyname name a

Orphan instances

PrettyClassicBy configName (tyname a) => PrettyBy (PrettyConfigClassic configName) (TyVarDecl tyname a)

Instance details

Methods

prettyBy :: PrettyConfigClassic configName -> TyVarDecl tyname a -> Doc ann

(PrettyClassicBy configName (tyname a), PrettyClassicBy configName (name a)) => PrettyBy (PrettyConfigClassic configName) (VarDecl tyname name a)

Instance details

Methods

prettyBy :: PrettyConfigClassic configName -> VarDecl tyname name a -> Doc ann

PrettyClassic (tyname a) => Pretty (TyVarDecl tyname a)

Instance details

Methods

pretty :: TyVarDecl tyname a -> Doc ann

prettyList :: [TyVarDecl tyname a] -> Doc ann

(PrettyClassic (tyname a), PrettyClassic (name a)) => Pretty (VarDecl tyname name a)

Instance details

Methods

pretty :: VarDecl tyname name a -> Doc ann

prettyList :: [VarDecl tyname name a] -> Doc ann