Language.PlutusIR.Compiler

Documentation

compileTerm :: Compiling m e a => Term TyName Name a -> m (PLCTerm a)

Compile a Term into a PLC Term. Note: the result *does* have globally unique names.

compileProgram :: Compiling m e a => Program TyName Name a -> m (Program TyName Name (Provenance a))

Compile a Program into a PLC Program. Note: the result *does* have globally unique names.

type Compiling m e a = (Monad m, MonadReader (Provenance a) m, AsError e (Provenance a), MonadError e m, MonadQuote m)

data Error a

Constructors

CompilationError a Text
A generic compilation error.

UnsupportedError a Text
An error relating specifically to an unsupported feature.

PLCError (Error a)
An error from running some PLC function, lifted into this error type for convenience.

class AsError r a | r -> a where

Minimal complete definition

_Error

Methods

_Error :: Prism' r (Error a)

_CompilationError :: Prism' r (a, Text)

_UnsupportedError :: Prism' r (a, Text)

_PLCError :: Prism' r (Error a)

data Provenance a

Indicates where a value comes from.

This is either an original annotation or a pieces of context explaining how the term relates to a previous Provenance. We also provide NoProvenance for convenience.

The provenance should always be just the original annotation, if we have one. It should only be another kind of provenance if we're in the process of generating some term that doesn't correspond directly to a term in the original AST.

Constructors

Original a

LetBinding Recursivity (Provenance a)

TermBinding String (Provenance a)

TypeBinding String (Provenance a)

DatatypeComponent DatatypeComponent (Provenance a)

NoProvenance