Language.PlutusIR.MkPir

Documentation

data Def var val

A definition. Pretty much just a pair with more descriptive names.

Constructors

Def

Fields

defVar :: var

defVal :: val

type TermDef tyname name a = Def (VarDecl tyname name a) (Term tyname name a)

A term definition as a variable.

type TypeDef tyname a = Def (TyVarDecl tyname a) (Type tyname a)

A type definition as a type variable.

type DatatypeDef tyname name a = Def (TyVarDecl tyname a) (Datatype tyname name a)

A datatype definition as a type variable.

mkVar :: a -> VarDecl tyname name a -> Term tyname name a

Make a Var referencing the given VarDecl.

mkTyVar :: a -> TyVarDecl tyname a -> Type tyname a

Make a TyVar referencing the given TyVarDecl.

mkLet :: a -> Recursivity -> [Binding tyname name a] -> Term tyname name a -> Term tyname name a

Make a let binding, unless the list of bindings is empty, in which case just return the underlying term.

mkIterTyForall :: a -> [TyVarDecl tyname a] -> Type tyname a -> Type tyname a

Universally quantify a list of names.

mkIterTyLam :: a -> [TyVarDecl tyname a] -> Type tyname a -> Type tyname a

Lambda abstract a list of names.

mkIterApp

Arguments

:: a

-> Term tyname name a
f

-> [Term tyname name a]
[ x0 ... xn ]

-> Term tyname name a
[f x0 ... xn ]

Make an iterated application.

mkIterTyFun :: a -> [Type tyname a] -> Type tyname a -> Type tyname a

Make an iterated function type.

mkIterLamAbs :: a -> [VarDecl tyname name a] -> Term tyname name a -> Term tyname name a

Lambda abstract a list of names.

mkIterInst

Arguments

:: a

-> Term tyname name a
a

-> [Type tyname a]
 [ x0 ... xn ]

-> Term tyname name a
{ a x0 ... xn }

Make an iterated instantiation.

mkIterTyAbs :: a -> [TyVarDecl tyname a] -> Term tyname name a -> Term tyname name a

Type abstract a list of names.

mkIterTyApp

Arguments

:: a

-> Type tyname a
f

-> [Type tyname a]
[ x0 ... xn ]

-> Type tyname a
[ f x0 ... xn ]

Make an iterated type application.