Language.PlutusIR.Optimizer.DeadCode

Description

Optimization passes for removing dead code, mainly dead let bindings.

Documentation

removeDeadBindings :: (HasUnique (name a) TermUnique, HasUnique (tyname a) TypeUnique) => Term tyname name a -> Term tyname name a

Remove all the dead let bindings in a term.